% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BASiCS.R
\name{basics_norm}
\alias{basics_norm}
\title{BASiCS normalisation}
\usage{
basics_norm(dCounts, sing_cols, outDir)
}
\arguments{
\item{dCounts}{The raw counts including ERCC spike in data}

\item{sing_cols}{The column names for the single cell data}
}
\description{
BASiCS normalisation
}

