% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{biotypes}
\alias{biotypes}
\title{Define the row colours for the heatmap.}
\usage{
biotypes(m)
}
\arguments{
\item{m}{A data frame with gene symbols as row names.}
}
\description{
Define the row colours for the heatmap.
}

