% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCluster.R
\name{bootstrapCluster}
\alias{bootstrapCluster}
\title{Assess cluster stability by bootstrapping}
\usage{
bootstrapCluster(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{bootstrapStability}}.}
}
\value{
See \code{\link{bootstrapStability}} for details.
}
\description{
Generate bootstrap replicates and recluster on them to determine the stability of clusters with respect to sampling noise.
This has been migrated to the \pkg{bluster} package as \code{\link{bootstrapStability}}.
}
\seealso{
\code{\link{bootstrapStability}}, the new name for this function.
}
\author{
Aaron Lun
}
