% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterModularity.R
\name{clusterModularity}
\alias{clusterModularity}
\title{Compute the cluster-wise modularity}
\usage{
clusterModularity(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{pairwiseModularity}}.}
}
\value{
See \code{?\link{pairwiseModularity}} for more details.
}
\description{
Calculate the modularity of each cluster from a graph, based on a null model of random connections between nodes.
This function has been moved to the \pkg{bluster} package as \code{\link{pairwiseModularity}}.
}
\seealso{
\code{\link{pairwiseModularity}}, the new name for this function.
}
\author{
Aaron Lun
}
