% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSilhouette.R
\name{clusterSilhouette}
\alias{clusterSilhouette}
\alias{clusterSilhouette,ANY-method}
\alias{clusterSilhouette,SummarizedExperiment-method}
\alias{clusterSilhouette,SingleCellExperiment-method}
\title{Approximate silhouette width}
\usage{
clusterSilhouette(x, ...)

\S4method{clusterSilhouette}{ANY}(x, ..., transposed = FALSE, subset.row = NULL)

\S4method{clusterSilhouette}{SummarizedExperiment}(x, ..., assay.type = "logcounts")

\S4method{clusterSilhouette}{SingleCellExperiment}(
  x,
  clusters = colLabels(x, onAbsence = "error"),
  ...,
  assay.type = "logcounts",
  use.dimred = NULL
)
}
\arguments{
\item{x}{A matrix-like object containing expression values for each gene (row) in each cell (column).
These dimensions can be transposed if \code{transposed=TRUE}.

Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such an expression matrix.
If \code{x} is a SingleCellExperiment and \code{use.dimred} is set, its \code{\link{reducedDims}} will be used instead.}

\item{...}{For the generic, arguments to pass to specific methods.

For the ANY method, arguments to pass to \code{\link{neighborPurity}}.

For the SummarizedExperiment method, arguments to pass to the ANY method.

For the SingleCellExperiment method, arguments to pass to the SummarizedExperiment method.}

\item{transposed}{A logical scalar indicating whether \code{x} is transposed (i.e., rows are cells).}

\item{subset.row}{See \code{?"\link{scran-gene-selection}"}.
Only used when \code{transposed=FALSE}.}

\item{assay.type}{A string specifying which assay values to use.}

\item{clusters}{A vector or factor of cluster IDs to pass to \code{\link{neighborPurity}}.}

\item{use.dimred}{A string specifying whether existing values in \code{reducedDims(x)} should be used.}
}
\value{
A \linkS4class{DataFrame} of purity statistics where each row corresponds to a cell in \code{x}, 
see \code{?\link{approxSilhouette}} for details.
}
\description{
\linkS4class{SingleCellExperiment}-compatible wrapper around \code{\link{approxSilhouette}},
for quickly computing an approximate silhouette width for each cell.
}
\examples{
library(scuttle)
sce <- mockSCE()
sce <- logNormCounts(sce)

g <- buildSNNGraph(sce)
clusters <- igraph::cluster_walktrap(g)$membership
out <- clusterSilhouette(sce, clusters)
boxplot(split(out$width, clusters))

# Mocking up a stronger example:
ngenes <- 1000
centers <- matrix(rnorm(ngenes*3), ncol=3)
clusters <- sample(1:3, ncol(sce), replace=TRUE)

y <- centers[,clusters]
y <- y + rnorm(length(y))

out2 <- clusterSilhouette(y, clusters)
boxplot(split(out2$width, clusters))

}
\seealso{
\code{\link{approxSilhouette}}, for the details of the approximate calculations.

\code{\link{clusterPurity}}, for another method of evaluating cluster separation.
}
\author{
Aaron Lun
}
