% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{trendVar}
\alias{decomposeVar}
\alias{testVar}
\alias{improvedCV2}
\alias{technicalCV2}
\alias{makeTechTrend}
\alias{multiBlockVar}
\alias{multiBlockNorm}
\alias{overlapExprs}
\alias{parallelPCA}
\title{Defunct functions}
\usage{
trendVar(...)

decomposeVar(...)

testVar(...)

improvedCV2(...)

technicalCV2(...)

makeTechTrend(...)

multiBlockVar(...)

multiBlockNorm(...)

overlapExprs(...)

parallelPCA(...)
}
\arguments{
\item{...}{Ignored arguments.}
}
\value{
All functions error out with a defunct message pointing towards its descendent (if available).
}
\description{
Functions that have passed on to the function afterlife.
Their successors are also listed.
}
\details{
\code{trendVar}, \code{decomposeVar} and \code{testVar} are succeeded by a suite of funtions related to \code{\link{modelGeneVar}} and \code{\link{fitTrendVar}}.

\code{improvedCV2} and \code{technicalCV2} are succeeded by \code{\link{modelGeneCV2}} and \code{\link{fitTrendCV2}}.

\code{makeTechTrend} is succeeded by \code{\link{modelGeneVarByPoisson}}.

\code{multiBlockVar} is succeeded by the \code{block} argument in many of the modelling functions, and \code{multiBlockNorm} is no longer necessary.

\code{\link{overlapExprs}} is succeeded by \code{\link{findMarkers}} with \code{test.type="wilcox"}.

\code{\link{parallelPCA}} has been moved over to the \pkg{PCAtools} package.
}
\examples{
try(trendVar())
}
\author{
Aaron Lun
}
