% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{fake_bulk}
\alias{fake_bulk}
\title{Add fake bulk sample column created from mean of all single cells}
\usage{
fake_bulk(dCounts)
}
\description{
Add fake bulk sample column created from mean of all single cells
}

