% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{filter}
\alias{filter}
\title{Filter the counts by CPM and percentage of samples}
\usage{
filter(dCounts, sing_cols, cpmVal, samp_pc)
}
\arguments{
\item{dCounts}{The counts dataframe}

\item{sing_cols}{A vector of the single cell column names}

\item{cpmVal}{The minimum CPM value}

\item{pc}{The percentage of cells to be expressed at the minumum CPM}
}
\value{
The original and filtered counts
}
\description{
Filter the counts by CPM and percentage of samples
}

