% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter the counts by CPM and percentage of samples}
\usage{
filter_counts(dCounts, sing_cols, cpmVal = 1, samp_pc = 5)
}
\arguments{
\item{dCounts}{The counts dataframe}

\item{sing_cols}{A vector of the single cell column names}

\item{cpmVal}{The minimum CPM value}

\item{pc}{The percentage of cells to be expressed at the minumum CPM}
}
\value{
The original and filtered counts
}
\description{
Filter the counts by CPM and percentage of samples
}

