% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSubCluster.R
\name{findSubCluster}
\alias{findSubCluster}
\alias{findSubCluster,ANY-method}
\alias{findSubCluster,SummarizedExperiment-method}
\alias{findSubCluster,SingleCellExperiment-method}
\title{Find subclusters under one or more cluster}
\usage{
findSubCluster(x, ...)

\S4method{findSubCluster}{ANY}(x, groups, normalize = TRUE, ...)

\S4method{findSubCluster}{SummarizedExperiment}(x, groups, ...)

\S4method{findSubCluster}{SingleCellExperiment}(x, groups = NULL, ...)
}
\arguments{
\item{x}{A matrix of counts or log-normalized expression values (if
\code{normalize=FALSE}), where each row corresponds to a gene and each column
corresponds to a cell.

Alternatively, a \linkS4class{SummarizedExperiment} or
\linkS4class{SingleCellExperiment} object containing such a matrix.}

\item{...}{For the generic, further arguments to pass to specific methods.

For the ANY and SummarizedExperiment methods, further arguments to pass to
the SingleCellExperiment method.}

\item{groups}{A vector of group assignments for all cells, usually
corresponding to cluster identities. If NULL, will determined by
\code{colLabels(x)}}

\item{normalize}{Logical scalar indicating whether each subset of \code{x}
should be log-transformed prior to further analysis.}

\item{clusters}{the clusters to be sub-clustered}

\item{prepFUN}{A function that accepts a single
\linkS4class{SingleCellExperiment} object and returns another
\linkS4class{SingleCellExperiment} containing any additional elements
required for clustering (e.g., PCA results). Using \code{dimred} to specify
the element names.}

\item{dimred}{String or integer scalar specifying the existing dimensionality reduction results to use..}

\item{BLUSPARAM}{A \linkS4class{BlusterParam} object that is used to specify
the clustering via \code{\link{clusterRows}}.}

\item{format}{A string to be passed to \code{\link{sprintf}}, specifying how
the subclusters should be named with respect to the parent level in
\code{groups} and the level returned by \code{clusterFUN}.}

\item{assay.type}{String or integer scalar specifying the relevant assay.}
}
\value{
a factor of length equal to ncol(x) containing cluster assignments
for each column of x.
}
\description{
Performs a quick subclustering for all cells under one or more cluster.
}
