% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTopCorrelations.R
\name{findTopCorrelations}
\alias{findTopCorrelations}
\alias{findTopCorrelations,ANY-method}
\alias{findTopCorrelations,SummarizedExperiment-method}
\title{Find top correlations between genes}
\usage{
findTopCorrelations(x, number, ...)

\S4method{findTopCorrelations}{ANY}(
  x,
  number = 10,
  y = NULL,
  d = 50,
  use.names = TRUE,
  BSPARAM = IrlbaParam(),
  BNPARAM = KmknnParam(),
  BPPARAM = SerialParam()
)

\S4method{findTopCorrelations}{SummarizedExperiment}(x, number, y = NULL, ..., assay.type = "logcounts")
}
\arguments{
\item{x}{A numeric matrix containing a (possibly log-transformed) normalized expression matrix for genes (rows) and cells (columns).
Alternatively, a \linkS4class{SummarizedExperiment} containing such a matrix.}

\item{number}{Integer scalar specifying the number of top correlated genes to report for each gene in \code{x}.}

\item{...}{For the generic, further arguments to pass to specific methods.

For the SummarizedExperiment method, further arguments to pass to the ANY method.}

\item{y}{Another normalized expression matrix with the same number of cells as \code{x}, possibly with different features.
For the SummarizedExperiment method, this may also be a SummarizedExperiment object containing such a matrix.}

\item{d}{Integer scalar specifying the number of dimensions to use for the approximate search via PCA.
If \code{NA}, no approximation of the rank values is performed prior to the search.}

\item{use.names}{Logical scalar specifying whether row names of \code{x} and/or \code{y} should be reported in the output, if available.}

\item{BSPARAM}{A \linkS4class{BiocSingularParam} object specifying the algorithm to use for the PCA.}

\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying the algorithm to use for the neighbor search.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying the parallelization scheme to use.}

\item{assay.type}{String or integer scalar specifying the assay containing the matrix of interest in \code{x} (and \code{y}, if a SummarizedExperiment).}
}
\value{
A \linkS4class{DataFrame} with up to \code{nrow(x) * number} rows, containing the top \code{number} correlated genes for each gene in \code{x}.
This contains the following fields:
\itemize{
\item \code{gene1}, the name (character) or row index (integer) of each gene in \code{x}.
\item \code{gene2}, the name (character) or row index (integer) of one of the top correlated genes to \code{gene1}.
This is another gene in \code{x} if \code{y=NULL}, otherwise it is a gene in \code{y}.
\item \code{rho}, the Spearman rank correlation for the current pair of \code{gene1} and \code{gene2}.
\item \code{p.value}, the approximate p-value associated with \code{rho} under the null hypothesis that the correlation is zero.
\item \code{FDR}, the adjusted p-value.
}
}
\description{
For each gene, find the subset of other genes that have strongest positive/negative Spearman's rank correlations in a normalized expression matrix.
}
\details{
In most cases, we only care about the top-correlated genes, allowing us to skip a lot of unnecessary computation.
This is achieved by transforming the problem of finding the largest Spearman correlation into a nearest-neighbor search in rank space.
For the sake of speed, we approximate the search by performing PCA to compress the rank values for all genes.

We compute the p-value for each gene using the approximate method implemented in \code{\link{cor.test}}.
The FDR correction is performed by considering all possible pairs of genes, as these are implicitly tested in the neighbor search.
Note that this is somewhat conservative as it does not consider strong correlations outside the reported genes.
}
\author{
Aaron Lun
}
