% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarkerEffects.R
\name{getMarkerEffects}
\alias{getMarkerEffects}
\title{Get marker effect sizes}
\usage{
getMarkerEffects(x, prefix = "logFC", strip = TRUE)
}
\arguments{
\item{x}{A \linkS4class{DataFrame} containing marker statistics for a given group/cluster,
usually one element of the List returned by \code{\link{findMarkers}}.}

\item{prefix}{String containing the prefix for the columns containing the effect size.}

\item{strip}{Logical scalar indicating whether the prefix should be removed from the output column names.}
}
\value{
A numeric matrix containing the effect sizes for the comparison to every other group/cluster.
}
\description{
Utility function to extract the marker effect sizes as a matrix from the output of \code{\link{findMarkers}}.
}
\examples{
library(scater)
sce <- mockSCE()
sce <- logNormCounts(sce)
                                                                     
kout <- kmeans(t(logcounts(sce)), centers=4) 
out <- findMarkers(sce, groups=kout$cluster)

eff1 <- getMarkerEffects(out[[1]])
str(eff1)
}
\seealso{
\code{\link{findMarkers}} and \code{\link{combineMarkers}}, to generate the DataFrames.
}
\author{
Aaron Lun
}
