% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopMarkers.R
\name{getTopMarkers}
\alias{getTopMarkers}
\title{Get top markers}
\usage{
getTopMarkers(de.lists, pairs, n = 10, field = "p.value",
  pairwise = TRUE, combine.type = c("any", "all"))
}
\arguments{
\item{de.lists}{A list-like object where each element is a data.frame or \linkS4class{DataFrame}.
Each element should represent the results of a pairwise comparison between two groups/clusters,
in which each row should contain the statistics for a single gene/feature.
Rows should be named by the feature name in the same order for all elements.}

\item{pairs}{A matrix, data.frame or \linkS4class{DataFrame} with two columns and number of rows equal to the length of \code{de.lists}.
Each row should specify the pair of clusters being compared for the corresponding element of \code{de.lists}.}

\item{n}{Integer scalar specifying the number of markers to obtain from each pairwise comparison.}

\item{field}{String specifying the column of each DataFrame in \code{de.lists} to use to identify top markers.
Smaller values are assigned higher rank.}

\item{pairwise}{Logical scalar indicating whether top markers should be returned for every pairwise comparison.
If \code{FALSE}, one marker set is returned for every cluster.}

\item{combine.type}{String specifying how markers from pairwise comparisons are to be combined if \code{pairwise=FALSE}.
If \code{"all"}, the intersection of pairwise marker sets is used, while if \code{"any"}, the union is used.}
}
\value{
If \code{pairwise=TRUE}, a \linkS4class{List} of Lists of character vectors is returned.
Each element of the outer list corresponds to cluster X, each element of the inner list corresponds to another cluster Y,
and each character vector specifies the marker genes that distinguish X from Y.

If \code{pairwise=FALSE}, a List of character vectors is returned.
Each character vector contains the marker genes that distinguish X from any/all other clusters.
}
\description{
Obtain the top markers for each pairwise comparison between clusters, or for each cluster.
}
\details{
This is a convenience utility that converts the results of pairwise comparisons into a marker list
that can be used in downstream functions, e.g., as the marker sets in \pkg{SingleR}.
}
\examples{
library(scater)
sce <- mockSCE()
sce <- logNormCounts(sce)

# Any clustering method is okay.
kout <- kmeans(t(logcounts(example.sce)), centers=3) 

out <- pairwiseTTests(logcounts(example.sce), 
     clusters=paste0("Cluster", kout$cluster))

# Getting top pairwise markers:
top <- getTopMarkers(out$statistics, out$pairs)
top[[1]]
top[[1]][[2]]

# Getting top per-cluster markers:
top <- getTopMarkers(out$statistics, out$pairs, pairwise=FALSE)
top[[1]]
}
\seealso{
\code{\link{pairwiseTTests}} and friends, to obtain \code{de.lists} and \code{pairs}.

\code{\link{combineMarkers}}, for another function that consolidates pairwise DE comparisons.
}
\author{
Aaron Lun
}
