% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulkDGE.R
\name{pseudoBulkDGE}
\alias{pseudoBulkDGE}
\alias{pseudoBulkDGE,ANY-method}
\alias{pseudoBulkDGE,SummarizedExperiment-method}
\title{Quickly perform pseudo-bulk DE analyses}
\usage{
pseudoBulkDGE(x, ...)

\S4method{pseudoBulkDGE}{ANY}(
  x,
  data,
  label,
  design,
  coef,
  contrast = NULL,
  condition = NULL,
  lfc = 0,
  include.intermediates = TRUE,
  sample = NULL
)

\S4method{pseudoBulkDGE}{SummarizedExperiment}(x, ..., assay.type = 1)
}
\arguments{
\item{x}{For the ANY method, a numeric matrix of counts where rows are genes and columns are pseudo-bulk profiles.

For the \linkS4class{SummarizedExperiment} method, a SummarizedExperiment object containing such a matrix in its assays.}

\item{...}{For the generic, additional arguments to pass to individual methods.

For the SummarizedExperiment method, additional arguments to pass to the ANY method.}

\item{data}{A data.frame or \linkS4class{DataFrame} containing metadata for each column of \code{x}.}

\item{label}{A vector of factor of length equal to \code{ncol(x)},
specifying the cluster or cell type assignment for each column of \code{x}.}

\item{design}{A formula to be used to construct a design matrix from variables in \code{data}.
Alternatively, a function that accepts a data.frame with the same fields as \code{data} and returns a design matrix.}

\item{coef}{String or character vector containing the coefficients to drop from the design matrix to form the null hypothesis.
Can also be an integer scalar or vector specifying the indices of the relevant columns.}

\item{contrast}{Numeric vector or matrix containing the contrast of interest.
Takes precedence over \code{coef}.}

\item{condition}{A vector or factor of length equal to \code{ncol(x)},
specifying the experimental condition for each column of \code{x}.
Only used for abundance-based filtering of genes.}

\item{lfc}{Numeric scalar specifying the log-fold change threshold to use in \code{\link{glmTreat}}.}

\item{include.intermediates}{Logical scalar indicating whether the intermediate \pkg{edgeR} objects should be returned.}

\item{sample}{Deprecated.}

\item{assay.type}{String or integer scalar specifying the assay to use from \code{x}.}
}
\value{
A \linkS4class{List} with one \linkS4class{DataFrame} of DE results per unique level of \code{cluster}.
This will contain at least the fields \code{"LogCPM"}, \code{"PValue"} and \code{"FDR"},
and usually \code{"logFC"} depending on whether an ANOVA-like contrast is requested in \code{coef} or \code{contrast}.
All DataFrames have row names equal to \code{rownames(x)}.

The \code{\link{metadata}} of the List contains \code{failed},
a character vector with the names of the labels for which the comparison could not be performed,
e.g., due to lack of residual d.f. 

If \code{include.intermediates}, the \code{\link{metadata}} of the individual DataFrames will also contain
\code{y}, the DGEList used for the analysis; and \code{fit}, the DGEGLM object after GLM fitting.
\code{fit} may not be reported if the corresponding label is in \code{failed}.
}
\description{
A wrapper function around \pkg{edgeR}'s quasi-likelihood methods
to conveniently perform differential expression analyses on pseudo-bulk profiles,
allowing detection of cell type-specific changes between conditions in replicated studies.
}
\details{
In replicated multi-condition scRNA-seq experiments,
we often have clusters comprised of cells from different samples of different experimental conditions.
It is often desirable to check for differential expression between conditions within each cluster,
allowing us to identify cell-type-specific responses to the experimental perturbation.

Given a set of pseudo-bulk profiles (usually generated by \code{\link{sumCountsAcrossCells}}),
this function loops over the labels and uses \pkg{edgeR} to detect DE genes between conditions.
The DE analysis for each label is largely the same as a standard analysis for bulk RNA-seq data,
using \code{design} and \code{coef} or \code{contrast} as described in the \pkg{edgeR} user's guide.
The analysis for each label is independent, i.e., the GLM fitting is performed separately,
to minimize problems due to differences in abundance and variance between labels.

Performing pseudo-bulk DGE enables us to re-use well-tested methods developed for bulk RNA-seq data analysis.
Each pseudo-bulk profile can be treated as an \emph{in silico} mimicry of a real bulk RNA-seq sample
(though in practice, it tends to be much more variable due to the relatively higher numbers of cells).
Pseudo-bulk analysis also allows direct modelling of variability between experimental replicates (i.e., across samples)
rather than that between cells in the same sample.
The former is more relevant to any statistical analysis that aims to obtain reproducible results.

In some cases, it will be impossible to perform a DE analysis for a label as there are no residual degrees of freedom.
This will be represented by a DataFrame with log-fold changes but \code{NA} p-values and FDRs.
In other cases, all statistics in the DataFrame will be \code{NA} if the contrast cannot be performed
(e.g., if a cell type only exists in one condition) or if the design matrix could not be constructed.

Note that we assume that \code{x} has already been filtered to remove unstable pseudo-bulk profiles generated from few cells.
}
\section{Comments on abundance filtering}{

For each label, abundance filtering is performed using \code{\link{filterByExpr}} prior to further analysis.
Genes that are filtered out will still show up in the DataFrame for that label, but with all statistics set to \code{NA}.
As this is done separately for each label, a different set of genes may be filtered out for each label,
which is largely to be expected if there is any label-specific expression.

By default, the minimum group size for \code{filterByExpr} is determined using the design matrix.
However, this may not be optimal if the design matrix contains additional terms (e.g., blocking factors)
in which case it is not easy to determine the minimum size of the groups relevant to the comparison of interest.
To overcome this, users can specify \code{condition.field} to specify the group to which each sample belongs,
which is used by \code{filterByExpr} to obtain a more appropriate minimum group size.
}

\examples{
set.seed(10000)
library(scuttle)
sce <- mockSCE(ncells=1000)
sce$samples <- gl(8, 125) # Pretending we have 8 samples.

# Making up some clusters.
sce <- logNormCounts(sce)
clusters <- kmeans(t(logcounts(sce)), centers=3)$cluster

# Creating a set of pseudo-bulk profiles:
info <- DataFrame(sample=sce$samples, cluster=clusters)
pseudo <- sumCountsAcrossCells(sce, info)

# Making up an experimental design for our 8 samples.
pseudo$DRUG <- gl(2,4)

# DGE analysis:
out <- pseudoBulkDGE(pseudo, 
   label=pseudo$cluster,
   condition=pseudo$DRUG,
   design=~DRUG,
   coef="DRUG2"
)
out[[1]]
metadata(out[[1]])
}
\references{
Tung P-Y et al. (2017).
Batch effects and the effective design of single-cell gene expression studies.
\emph{Sci. Rep.} 7, 39921

Lun ATL and Marioni JC (2017).
Overcoming confounding plate effects in differential expression analyses of single-cell RNA-seq data.
\emph{Biostatistics} 18, 451-464

Crowell HL et al. (2019).
On the discovery of population-specific state transitions from multi-sample multi-condition single-cell RNA sequencing data.
\emph{biorXiv}
}
\seealso{
\code{\link{sumCountsAcrossCells}}, to easily generate the pseudo-bulk count matrix.

\code{\link{decideTestsPerLabel}}, to generate a summary of the DE results across all labels.

\code{pbDS} from the \pkg{muscat} package, which uses a similar approach.
}
\author{
Aaron Lun
}
