% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulkSpecific.R
\name{pseudoBulkSpecific}
\alias{pseudoBulkSpecific}
\alias{pseudoBulkSpecific,ANY-method}
\alias{pseudoBulkSpecific,SummarizedExperiment-method}
\title{Label-specific pseudo-bulk DE}
\usage{
pseudoBulkSpecific(x, ...)

\S4method{pseudoBulkSpecific}{ANY}(
  x,
  label,
  condition,
  ...,
  method = c("edgeR", "voom"),
  sorted = FALSE,
  average = c("median", "mean"),
  reference = NULL
)

\S4method{pseudoBulkSpecific}{SummarizedExperiment}(x, ..., assay.type = 1)
}
\arguments{
\item{x}{A numeric matrix of counts where rows are genes and columns are pseudo-bulk profiles.
Alternatively, a SummarizedExperiment object containing such a matrix in its assays.}

\item{...}{For the generic, further arguments to pass to individual methods.

For the ANY method, further arguments to pass to \code{\link{pseudoBulkDGE}}.

For the SummarizedExperiment method, further arguments to pass to the ANY method.}

\item{label}{A vector of factor of length equal to \code{ncol(x)},
specifying the cluster or cell type assignment for each column of \code{x}.}

\item{condition}{A vector or factor of length equal to \code{ncol(x)},
specifying the experimental condition for each column of \code{x}.
Only used for abundance-based filtering of genes.}

\item{method}{String specifying the DE analysis framework to use.}

\item{sorted}{Logical scalar indicating whether the output tables should be sorted by p-value.}

\item{average}{String specifying the method to compute the average log-fold change of all other labels.}

\item{reference}{A List containing the (unsorted) output of \code{\link{pseudoBulkDGE}}.
This can be supplied to avoid redundant calculations but is automatically computed if \code{NULL}.}

\item{assay.type}{String or integer scalar specifying the assay to use from \code{x}.}
}
\value{
A \linkS4class{List} of \linkS4class{DataFrame}s where each DataFrame contains DE statistics for one label.
This is equivalent to the output of \code{\link{pseudoBulkDGE}};
if \code{reference} is supplied, most of the statistics will be identical to those reported there.

The main differences are that the p-values and FDR values are changed.
Each DataFrame also has an \code{OtherAverage} field containing the average log-fold change across all other labels.
}
\description{
Detect label-specific DE genes in a pseudo-bulk analysis,
by testing whether the log-fold change is more extreme than the average log-fold change of other labels.
}
\details{
This function implements a quick and dirty method for detecting label-specific DE genes.
For a given label and gene, the null hypothesis is that the log-fold change lies between zero
and the average log-fold change for that gene across all other labels.
This is tested by testing each gene against the two extremes and taking the larger of the two p-values, 
as well as setting the p-value to 1 if the log-fold change lies between the extremes.
}
\examples{
set.seed(10000)
library(scuttle)
sce <- mockSCE(ncells=1000)
sce$samples <- gl(8, 125) # Pretending we have 8 samples.

# Making up some clusters.
sce <- logNormCounts(sce)
clusters <- kmeans(t(logcounts(sce)), centers=3)$cluster

# Creating a set of pseudo-bulk profiles:
info <- DataFrame(sample=sce$samples, cluster=clusters)
pseudo <- sumCountsAcrossCells(sce, info)

# Making up an experimental design for our 8 samples
# and adding a common DE across all labels.
pseudo$DRUG <- gl(2,4)[pseudo$sample]
assay(pseudo)[1,pseudo$DRUG==1] <- assay(pseudo)[1,pseudo$DRUG==1] * 10 

# Label-specific analysis:
out <- pseudoBulkSpecific(pseudo, 
   label=pseudo$cluster,
   condition=pseudo$DRUG,
   design=~DRUG,
   coef="DRUG2"
)

out[[1]]

}
\seealso{
\code{\link{pseudoBulkDGE}}, for the underlying function that does all the heavy lifting.
}
\author{
Aaron Lun
}
