% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{sing_col_convert}
\alias{sing_col_convert}
\title{Convert single cell column numbers to names}
\usage{
sing_col_convert(dCounts, sing_cols)
}
\description{
Convert single cell column numbers to names
}

