% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{biotypes}
\alias{biotypes}
\title{Assign biotypes to each gene}
\usage{
biotypes(geneCounts, species)

biotypes(geneCounts, species)
}
\arguments{
\item{m}{A data frame with gene symbols as row names.}
}
\description{
Assign biotypes to each gene

Biotypes
}

