% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brennecke.R
\name{brennecke}
\alias{brennecke}
\title{Requires df of gene symbols, lengths and counts, rownames =  ensembl IDs and colnames = single cell IDs}
\usage{
brennecke(dCounts, species, outDir, spike_text)
}
\description{
Requires df of gene symbols, lengths and counts, rownames =  ensembl IDs and colnames = single cell IDs
}

