% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSilhouette.R
\name{clusterSilhouette}
\alias{clusterSilhouette}
\alias{clusterSilhouette,ANY-method}
\alias{clusterSilhouette,SummarizedExperiment-method}
\alias{clusterSilhouette,SingleCellExperiment-method}
\title{Approximate silhouette width}
\usage{
clusterSilhouette(x, ...)

\S4method{clusterSilhouette}{ANY}(
  x,
  clusters,
  k = 50,
  transposed = FALSE,
  subset.row = NULL,
  BPPARAM = SerialParam()
)

\S4method{clusterSilhouette}{SummarizedExperiment}(x, ..., assay.type = "logcounts")

\S4method{clusterSilhouette}{SingleCellExperiment}(
  x,
  clusters = colLabels(x, onAbsence = "error"),
  ...,
  assay.type = "logcounts",
  use.dimred = NULL
)
}
\arguments{
\item{x}{A numeric matrix-like object containing expression values for genes (rows) and cells (columns).
If \code{transposed=TRUE}, cells should be in rows and reduced dimensions should be in the columns.

Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} object containing such a matrix.

For the SingleCellExperiment method, if \code{use.dimred} is supplied, 
\code{x} contain an appropriate reduced dimension result in its \code{\link{reducedDims}}.}

\item{...}{For the generic, arguments to pass to specific methods.

For the SummarizedExperiment method, arguments to pass to the ANY method.

For the SingleCellExperiment method, arguments to pass to the SummarizedExperiment method.}

\item{clusters}{Vector of length equal to \code{ncol(x)}, specifying the cluster identity for each cell.}

\item{k}{Integer scalar specifying the number of nearest neighbors to use to determine the radius of the hyperspheres.}

\item{transposed}{Logical scalar specifying whether \code{x} contains cells in the rows.}

\item{subset.row}{See \code{?"\link{scran-gene-selection}"}.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating whether and how parallelization should be performed across genes.}

\item{assay.type}{A string specifying which assay values to use.}

\item{use.dimred}{A string specifying whether existing values in \code{reducedDims(x)} should be used.}
}
\value{
A \linkS4class{DataFrame} with one row per cell in \code{x} and the columns:
\itemize{
\item \code{width}, a numeric field containing the approximate silhouette width of the current cell.
\item \code{other}, the closest cluster other than the one to which the current cell is assigned.
}
Row names are defined as the column names of \code{x}.
}
\description{
Given a clustering, compute a fast approximate silhouette width for each cell.
}
\details{
The silhouette width is a general purpose method for evaluating the separation between clusters 
but requires calculating the average distance between pairs of cells within or between clusters.
This function approximates the average distance with \eqn{d}, defined as the larger of:
\itemize{
\item The distance from the current cell to the centroid of cluster \eqn{X}.
This is a reasonable approximation when the cell is distant to \eqn{X} relative to the latter's variation.
\item The root mean squared distance between cells in cluster \eqn{X}.
This is a reasonable approximation when the cell lies inside the bulk of points for \eqn{X}.
}
The approximate silhouette width for each cell can then be calculated with two values of \eqn{d},
computed by setting \eqn{X} to the cluster of the current cell or the closest other cluster.
}
\examples{
library(scuttle)
sce <- mockSCE()
sce <- logNormCounts(sce)

g <- buildSNNGraph(sce)
clusters <- igraph::cluster_walktrap(g)$membership
out <- clusterSilhouette(sce, clusters)
boxplot(split(out$width, clusters))

# Mocking up a stronger example:
ngenes <- 1000
centers <- matrix(rnorm(ngenes*3), ncol=3)
clusters <- sample(1:3, ncol(sce), replace=TRUE)

y <- centers[,clusters]
y <- y + rnorm(length(y))

out2 <- clusterSilhouette(y, clusters)
boxplot(split(out2$width, clusters))

}
\seealso{
\code{silhouette} from the \pkg{cluster} package, for the exact calculation.

\code{\link{clusterPurity}}, for another method of evaluating cluster separation.
}
\author{
Aaron Lun
}
