% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSumFactors.R
\name{computeSumFactors}
\alias{computeSumFactors}
\alias{calculateSumFactors}
\title{Normalization by deconvolution}
\usage{
computeSumFactors(...)

calculateSumFactors(...)
}
\arguments{
\item{...}{Further arguments to pass to \code{\link{pooledSizeFactors}} or \code{\link{computePooledFactors}}.}
}
\value{
For \code{calculateSumFactors}, a numeric vector of size factors returned by \code{\link{pooledSizeFactors}}.

For \code{computeSumFactors}, a SingleCellExperiment containing the size factors in its \code{\link{sizeFactors}},
as returned by \code{\link{computePooledFactors}}.
}
\description{
Scaling normalization of single-cell RNA-seq data by deconvolving size factors from cell pools.
These functions have been moved to the \pkg{scuttle} package and are just retained here for compatibility.
}
\author{
Aaron Lun
}
