% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSubCluster.R
\name{.find_sub_cluster}
\alias{.find_sub_cluster}
\alias{findSubCluster}
\alias{findSubCluster,ANY-method}
\alias{findSubCluster,SummarizedExperiment-method}
\alias{findSubCluster,SingleCellExperiment-method}
\title{Find subclusters under one or more cluster}
\usage{
.find_sub_cluster(
  x,
  groups = NULL,
  clusters = NULL,
  normalize = TRUE,
  prepFUN = NULL,
  clusterFUN = NULL,
  BLUSPARAM = NNGraphParam(),
  format = "\%s.\%s",
  assay.type = NULL
)

findSubCluster(x, ...)

\S4method{findSubCluster}{ANY}(x, normalize = TRUE, ...)

\S4method{findSubCluster}{SummarizedExperiment}(x, ...)

\S4method{findSubCluster}{SingleCellExperiment}(
  x,
  groups = NULL,
  clusters = NULL,
  normalize = TRUE,
  prepFUN = NULL,
  clusterFUN = NULL,
  BLUSPARAM = NNGraphParam(),
  format = "\%s.\%s",
  assay.type = NULL
)
}
\arguments{
\item{x}{A matrix of counts or log-normalized expression values (if
\code{normalize=FALSE}), where each row corresponds to a gene and each column
corresponds to a cell.

Alternatively, a \linkS4class{SummarizedExperiment} or
\linkS4class{SingleCellExperiment} object containing such a matrix.}

\item{groups}{A vector of group assignments for all cells, usually
corresponding to cluster identities. If NULL, will determined by
\code{colLabels(x)}}

\item{clusters}{the cluster to be sub-clustered}

\item{normalize}{Logical scalar indicating whether each subset of \code{x}
should be log-transformed prior to further analysis.}

\item{prepFUN}{A function that accepts a single
\linkS4class{SingleCellExperiment} object and returns another
\linkS4class{SingleCellExperiment} containing any additional elements
required for clustering (e.g., PCA results).}

\item{clusterFUN}{A function that accepts a single
\linkS4class{SingleCellExperiment} object and returns a vector of cluster
assignments for each cell in that object.}

\item{BLUSPARAM}{A \linkS4class{BlusterParam} object that is used to specify
the clustering via \code{\link{clusterRows}}.  Only used when
\code{clusterFUN=NULL}.}

\item{format}{A string to be passed to \code{\link{sprintf}}, specifying how
the subclusters should be named with respect to the parent level in
\code{groups} and the level returned by \code{clusterFUN}.}

\item{assay.type}{String or integer scalar specifying the relevant assay.}

\item{...}{For the generic, further arguments to pass to specific methods.

For the ANY and SummarizedExperiment methods, further arguments to pass to
the SingleCellExperiment method.}
}
\value{
a factor of length equal to ncol(x) containing cluster assignments
for each column of x.
}
\description{
Performs a quick subclustering for all cells under one or more cluster.
}
