\name{Quick clustering}
\alias{quickCluster}
\alias{quickCluster,ANY-method}
\alias{quickCluster,SingleCellExperiment-method}

\title{Quick clustering of cells}
\description{Cluster similar cells based on their expression profiles, using either log-expression values or ranks.}

\usage{
\S4method{quickCluster}{ANY}(x, min.size=100, method=c("igraph", "hclust"), 
    use.ranks=NULL, d=NULL, subset.row=NULL, min.mean=1, 
    graph.fun=cluster_walktrap, BSPARAM=ExactParam(), BPPARAM=SerialParam(),
    block=NULL, block.BPPARAM=SerialParam(), ...)

\S4method{quickCluster}{SingleCellExperiment}(x, subset.row=NULL, ..., assay.type="counts", get.spikes=FALSE)
}

\arguments{
\item{x}{
    A numeric count matrix where rows are genes and columns are cells.
    Alternatively, a \linkS4class{SingleCellExperiment} object containing such a matrix.
}
\item{min.size}{An integer scalar specifying the minimum size of each cluster.}
\item{method}{A string specifying the clustering method to use.}
\item{use.ranks}{A logical scalar indicating whether clustering should be performed on the rank matrix, i.e., based on Spearman's rank correlation.
Defaults to \code{TRUE} for consistency with old defaults, but this will be changed to \code{FALSE} in subsequent releases.}
\item{d}{An integer scalar specifying the number of principal components to retain.

If \code{NULL} and \code{use.ranks=TRUE}, this defaults to 50.
If \code{use.rank=FALSE}, the number of PCs is chosen by \code{\link{denoisePCA}}.

If \code{NA}, no dimensionality reduction is performed and the gene expression values (or their rank equivalents) are directly used in clustering.
}
\item{subset.row}{See \code{?"\link{scran-gene-selection}"}.}
\item{min.mean}{A numeric scalar specifying the filter to be applied on the average count for each filter prior to computing ranks.
Only used when \code{use.ranks=TRUE}, see \code{?\link{scaledColRanks}} for details.}
\item{graph.fun}{A function specifying the community detection algorithm to use on the nearest neighbor graph when \code{method="igraph"}.
Usually obtained from the \pkg{igraph} package.}
\item{BSPARAM}{A \linkS4class{BiocSingularParam} object specifying the algorithm to use for PCA, if \code{d} is not \code{NA}.}
\item{BPPARAM}{A \linkS4class{BiocParallelParam} object to use for parallel processing within each block.}
\item{block}{A factor of length equal to \code{ncol(x)} specifying whether clustering should be performed within pre-specified blocks.
By default, all columns in \code{x} are treated as a single block.}
\item{block.BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether and how parallelization should be performed across blocks, 
if \code{block} is non-\code{NULL} and has more than one level.}
\item{...}{
    For \code{quickCluster,ANY-method}, additional arguments to be passed to \code{\link{cutreeDynamic}} for \code{method="hclust"}, 
        or \code{\link{buildSNNGraph}} for \code{method="igraph"}.

    For \code{quickCluster,SingleCellExperiment-method}, additional arguments to pass to \code{quickCluster,ANY-method}.
}
\item{assay.type}{A string specifying which assay values to use, e.g., \code{"counts"} or \code{"logcounts"}.}
\item{get.spikes}{See \code{?"\link{scran-gene-selection}"}.}
}

\details{
This function provides a convenient wrapper to quickly define clusters of a minimum size \code{min.size}.
Two clustering strategies are available:
\itemize{
\item If \code{method="hclust"}, a distance matrix is constructed;
hierarchical clustering is performed using Ward's criterion;
and \code{\link{cutreeDynamic}} is used to define clusters of cells.
\item If \code{method="igraph"}, a shared nearest neighbor graph is constructed using the \code{\link{buildSNNGraph}} function.
This is used to define clusters based on highly connected communities in the graph, using the \code{graph.fun} function.
}

By default, \code{quickCluster} will apply these clustering algorithms on the principal component (PC) scores generated from the log-expression values.
These are obtained by \code{\link{denoisePCA}} based on the trend fitted to endogenous genes with \code{\link{trendVar}}.

If \code{use.ranks=TRUE}, clustering is instead performed on PC scores obtained from scaled and centred ranks generated by \code{\link{scaledColRanks}}.
This effectively means that clustering uses distances based on the Spearman's rank correlation between two cells.
In addition, if \code{x} is a \linkS4class{dgCMatrix} and \code{BSPARAM} has \code{deferred=TRUE}, 
ranks will be computed without loss of sparsity to improve speed and memory efficiency during PCA.

Setting \code{use.ranks=TRUE} is invariant to scaling normalization and avoids circularity between normalization and clustering, e.g., in \code{\link{computeSumFactors}}.
However, the default is to use the log-expression values as this yields finer and more precise clusters.
}

\section{Enforcing cluster sizes}{
With \code{method="hclust"}, \code{\link{cutreeDynamic}} is used to ensure that all clusters contain a minimum number of cells.
However, some cells may not be assigned to any cluster and are assigned identities of \code{"0"} in the output vector.
In most cases, this is because those cells belong in a separate cluster with fewer than \code{min.size} cells.
The function will not be able to call this as a cluster as the minimum threshold on the number of cells has not been passed.
Users are advised to check that the unassigned cells do indeed form their own cluster.
Otherwise, it may be necessary to use a different clustering algorithm.

When using \code{method="igraph"}, clusters are first identified using the specified \code{graph.fun}.
If the smallest cluster contains fewer cells than \code{min.size}, it is merged with the closest neighbouring cluster.
In particular, the function will attempt to merge the smallest cluster with each other cluster.
The merge that maximizes the modularity score is selected, and a new merged cluster is formed.
This process is repeated until all (merged) clusters are larger than \code{min.size}.
}

\section{Gene selection}{
Spike-in transcripts are not used by default as they provide little information on the biological similarities between cells.
This may not be the case if subpopulations differ by total RNA content, in which case setting \code{get.spikes=TRUE} may provide more discriminative power.

When \code{use.ranks=TRUE}, the function will also filter out genes with average counts (as defined by \code{\link{calcAverage}}) below \code{min.mean}.
This removes low-abundance genes with many tied ranks, especially due to zeros, which may reduce the precision of the clustering.
We suggest setting \code{min.mean} to 1 for read count data and 0.1 for UMI data.
}

\value{
A character vector of cluster identities for each cell in \code{counts} is returned.
}

\author{
Aaron Lun and Karsten Bach
}

\seealso{
\code{\link{cutreeDynamic}},
\code{\link{computeSumFactors}},
\code{\link{buildSNNGraph}}

\code{\link{scaledColRanks}} to get the rank matrix directly.
}

\examples{
set.seed(100)
popsize <- 200
ngenes <- 1000
all.facs <- 2^rnorm(popsize, sd=0.5)
counts <- matrix(rnbinom(ngenes*popsize, mu=all.facs, size=1), ncol=popsize, byrow=TRUE)

clusters <- quickCluster(counts, use.ranks=FALSE, min.size=20)
clusters <- quickCluster(counts, use.ranks=TRUE)
}

\keyword{normalization}

\references{
van Dongen S and Enright AJ (2012).
Metric distances derived from cosine similarity and Pearson and Spearman correlations.
\emph{arXiv} 1208.3145

Lun ATL, Bach K and Marioni JC (2016).
Pooling across cells to normalize single-cell RNA sequencing data with many zero counts.
\emph{Genome Biol.} 17:75
}
